+++ demo/pug/invoice.pug
@@ -1,6 +1,6 @@
 html
   body
-    h1 Invoice for #{company}
+    h1 Pug: Invoice for #{company}
     p This is some text
     table
       thead
