+++ demo/twig/invoice.scss
@@ -1,0 +1,13 @@
+html, body {
+  margin: 16px;
+  padding: 0;
+}
+
+table {
+  padding: 8px;
+
+  th {
+    padding: 8px;
+    border: 1px solid #333;
+  }
+}
