+++ package.json
@@ -17,6 +17,7 @@
     "js-yaml": "^3.12.0",
     "node-sass": "^4.9.4",
     "pug": "^2.0.3",
-    "puppeteer": "^1.9.0"
+    "puppeteer": "^1.9.0",
+    "twig": "^1.12.0"
   }
 }
